% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe_1.R
\docType{data}
\name{anscombe_1}
\alias{anscombe_1}
\title{Modified Anscombe's quartet set 1}
\format{A data frame with 11 rows and 2 variables.
\describe{
\item{x_1, y}{Point coordinates}
}}
\source{
Anscombe FJ. Graphs in statistical analysis. The American
Statistician. 1973 Feb;27(1):17-21.
\url{https://www.jstor.org/stable/2682899}
}
\usage{
anscombe_1
}
\description{
Anscombe's quartet set 1 slightly modified so that ordinary least squares fit
has an intercept and slope of exactly 3 and 0.5, respectively.
}
\keyword{datasets}
