% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe_1_outlier.R
\docType{data}
\name{anscombe_1_outlier}
\alias{anscombe_1_outlier}
\title{Modified Anscombe's Quartet Set 1 with outlier}
\format{A data frame with 12 rows and 2 variables.
\describe{
\item{x_1, y}{Point coordinates}
}}
\source{
Anscombe FJ. Graphs in statistical analysis. The American
Statistician. 1973 Feb;27(1):17-21.
\url{https://www.jstor.org/stable/2682899}
}
\usage{
anscombe_1_outlier
}
\description{
The \code{anscombe_1} dataset with an added outlier.
}
\keyword{datasets}
