% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_pressure.R
\docType{data}
\name{blood_pressure}
\alias{blood_pressure}
\title{National Health and Nutrition Examination Survey blood pressure data.}
\format{
A data frame with 6144 rows and 5 variables.
\describe{
\item{seqn}{Respondent sequence number}
\item{ridageyr}{Age in years at screening}
\item{bpxosy1}{Systolic blood pressure - 1st oscillometric reading}
\item{bpxosy2}{Systolic blood pressure - 2nd oscillometric reading}
\item{bpxosy3}{Systolic blood pressure - 3rd oscillometric reading}
}
}
\source{
National Health and Nutrition Examination Survey 2017 - 2018.
\url{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?BeginYear=2017}
}
\usage{
blood_pressure
}
\description{
Systolic blood pressure data for 6144 participants.
}
\keyword{datasets}
