% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildings.R
\docType{data}
\name{buildings}
\alias{buildings}
\title{Building height data}
\format{
A data frame with 76 rows and 6 variables.
\describe{
\item{building}{Name of the building}
\item{city}{Home city of the building}
\item{country}{Home country of the building}
\item{height}{Height of the building, in feet}
\item{floors}{Number of floors}
\item{year}{Year the building was built}
}
}
\source{
Wikipedia.
\url{https://en.wikipedia.org/wiki/List_of_tallest_buildings#Tallest_buildings_in_the_world}
}
\usage{
buildings
}
\description{
Heights and other data on the world's 76 tallest buildings.
}
\keyword{datasets}
