% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress.R
\docType{data}
\name{congress}
\alias{congress}
\title{Data on 2019 Congress members}
\format{
A data frame with 532 rows and 7 variables.
\describe{
\item{name}{Congress member name}
\item{state}{State}
\item{division}{US Census division}
\item{chamber}{Chamber (House or Senate)}
\item{party}{Party}
\item{age}{Age on 2019-01-08}
\item{gender}{Gender}
}
}
\source{
The @unitedstates project.
\url{https://theunitedstates.io/}
}
\usage{
congress
}
\description{
Demographic and party data on 2019 House and Senate members. The data reflects
the members of Congress as of 2019-01-08.
}
\keyword{datasets}
