% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_acs_1.R
\docType{data}
\name{example_acs_1}
\alias{example_acs_1}
\title{Example 5-year ACS data, Set 1}
\format{A data frame with 156 rows and 4 variables.
\describe{
\item{geoid}{State FIPS}
\item{name}{State}
\item{variable}{ACS variable}
\item{estimate}{ACS variable estimate}
}}
\source{
U.S. Census Bureau. American Community Survey.
\url{https://www.census.gov/programs-surveys/acs}
}
\usage{
example_acs_1
}
\description{
Example 5-year ACS data, Set 1.
}
\keyword{datasets}
