% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_acs_4.R
\docType{data}
\name{example_acs_4}
\alias{example_acs_4}
\title{Example 5-year ACS data, Set 4}
\format{A data frame with 156 rows and 4 variables.
\describe{
\item{geoid}{State FIPS}
\item{name}{State}
\item{variable}{ACS variable}
\item{estimate}{ACS variable estimate}
\item{error}{ACS variable estimate error}
}}
\source{
U.S. Census Bureau. American Community Survey.
\url{https://www.census.gov/programs-surveys/acs}
}
\usage{
example_acs_4
}
\description{
Example 5-year ACS data, Set 4.
}
\keyword{datasets}
