% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_eagle_nests.R
\docType{data}
\name{example_eagle_nests}
\alias{example_eagle_nests}
\title{Example eagle nest data, not tidy}
\format{
A data frame with 3 rows and 3 variables.
\describe{
\item{region}{Region}
\item{2007, 2009}{Years}
}
}
\source{
U.S. Fish and Wildlife Service. Bald and Golden Eagles. 2016. p. 7.
\url{https://www.fws.gov/migratorybirds/pdf/management/EagleRuleRevisions-StatusReport.pdf}
}
\usage{
example_eagle_nests
}
\description{
Example eagle nest data, not tidy.
}
\keyword{datasets}
