% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_eagle_nests_tidy.R
\docType{data}
\name{example_eagle_nests_tidy}
\alias{example_eagle_nests_tidy}
\title{Example eagle nest data, tidy}
\format{
A data frame with 6 rows and 3 variables.
\describe{
\item{region}{Region}
\item{year}{Year}
\item{num_nests}{Number of nests observed}
}
}
\source{
U.S. Fish and Wildlife Service. Bald and Golden Eagles. 2016. p. 7.
\url{https://www.fws.gov/migratorybirds/pdf/management/EagleRuleRevisions-StatusReport.pdf}
}
\usage{
example_eagle_nests_tidy
}
\description{
Example eagle nest data, tidy.
}
\keyword{datasets}
