% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_eagle_pairs.R
\docType{data}
\name{example_eagle_pairs}
\alias{example_eagle_pairs}
\title{Example eagle pair data}
\format{
A data frame with 48 rows and 12 variables. The values are
observations of pairs of eagles.
\describe{
\item{state}{State}
\item{state_abbr}{State abbreviation}
\item{1997 - 2006}{Years}
}
}
\source{
U.S. Fish and Wildlife Service. Midwest Region - Bald and Golden
Eagles.
\url{https://www.fws.gov/midwest/eagle/NestingData/nos_state_tbl.html}
}
\usage{
example_eagle_pairs
}
\description{
Example eagle pair data.
}
\keyword{datasets}
