% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_gymnastics_1.R
\docType{data}
\name{example_gymnastics_1}
\alias{example_gymnastics_1}
\title{Example gymnastics data, set 1}
\format{A data frame with 3 rows and 3 variables.
\describe{
\item{country}{Country}
\item{score_vault}{Vault score}
\item{score_floor}{Floor score}
}}
\source{
Wikipedia.
}
\usage{
example_gymnastics_1
}
\description{
Women's gymnastics team scores in 2016 Olympics for the United States,
Russia, and China in the vault and floor events.
}
\keyword{datasets}
