% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_gymnastics_3.R
\docType{data}
\name{example_gymnastics_3}
\alias{example_gymnastics_3}
\title{Example gymnastics data, set 3}
\format{A data frame with 3 rows and 9 variables.
\describe{
\item{country}{Country}
\item{vault_2012_f}{Women's vault score in 2012 Olympics}
\item{vault_2012_m}{Men's vault score in 2012 Olympics}
\item{vault_2016_f}{Women's vault score in 2016 Olympics}
\item{vault_2016_m}{Men's vault score in 2016 Olympics}
\item{floor_2012_f}{Women's floor score in 2012 Olympics}
\item{floor_2012_m}{Men's floor score in 2012 Olympics}
\item{floor_2016_f}{Women's floor score in 2016 Olympics}
\item{floor_2016_m}{Men's floor score in 2016 Olympics}
}}
\source{
Wikipedia.
}
\usage{
example_gymnastics_3
}
\description{
Men's and women's gymnastics team scores in 2016 Olympics for the
United States, Russia, and China in the vault and floor events.
}
\keyword{datasets}
