% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_gymnastics_4.R
\docType{data}
\name{example_gymnastics_4}
\alias{example_gymnastics_4}
\title{Example gymnastics data, set 4}
\format{
A data frame with 3 rows and 5 variables.
\describe{
\item{country}{Country}
\item{vault2012}{Vault score in 2012 Olympics}
\item{floor2012}{Floor score in 2012 Olympics}
\item{vault2016}{Vault score in 2016 Olympics}
\item{floor2016}{Floor score in 2016 Olympics}
}
}
\source{
Wikipedia.
}
\usage{
example_gymnastics_4
}
\description{
Women's gymnastics team scores in the 2012 and 2016 Olympics for the
United States, Russia, and China in the vault and floor events.
}
\keyword{datasets}
