% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_gymnastics_5.R
\docType{data}
\name{example_gymnastics_5}
\alias{example_gymnastics_5}
\title{Example gymnastics data, set 5}
\format{
A data frame with 12 rows and 3 variables.
\describe{
\item{country}{Country}
\item{variable}{Indicator for total or rank and year}
\item{value}{Value of variable}
}
}
\source{
Wikipedia.
}
\usage{
example_gymnastics_5
}
\description{
Women's gymnastics team scores and ranks in the 2012 and 2016 Olympics for
the United States, Russia, and China.
}
\keyword{datasets}
