% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_migration.R
\docType{data}
\name{example_migration}
\alias{example_migration}
\title{Example migration data.}
\format{A data frame with 3 rows and 6 variables.
\describe{
\item{dest}{Destination country}
\item{Afghanistan}{Migrant population from Afghanistan}
\item{Canada}{Migrant population from Canada}
\item{India}{Migrant population from India}
\item{Japan}{Migrant population from Japan}
\item{South Africa}{Migrant population from South Africa}
}}
\source{
United Nations, Population Division, Department of Economic and
Social Affairs. Trends in International Migrant Stock, 2017. Table 1.
\url{https://www.un.org/en/development/desa/population/migration/data/estimates2/data/UN_MigrantStockByOriginAndDestination_2017.xlsx}
}
\usage{
example_migration
}
\description{
Migrant populations in Albania, Bulgaria, and Romania from Afghanistan
Canada, India, Japan, South Africa in 2017.
}
\keyword{datasets}
