% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sprint.R
\docType{data}
\name{example_sprint}
\alias{example_sprint}
\title{Example men's 100m dash data}
\format{A data frame with 11 rows and 2 variables.
\describe{
\item{x_1, y}{Point coordinates}
}}
\source{
Wikipedia.
\url{https://en.wikipedia.org/wiki/Athletics_at_the_2016_Summer_Olympics_\%E2\%80\%93_Men\%27s_100_metres#Final}
}
\usage{
example_sprint
}
\description{
Results of the men's 100-meter dash final from the 2016 Summer Olympics
}
\keyword{datasets}
