% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famines.R
\docType{data}
\name{famines}
\alias{famines}
\title{World famines data from Our World in Data}
\format{
A data frame with 77 rows and 6 variables.
\describe{
\item{location}{Famine location, typically a single country}
\item{iso_a3}{ISO code of famine location}
\item{region}{Region}
\item{year_start}{Year the famine began}
\item{year_end}{Year the famine ended}
\item{deaths_estimate}{Estimated mortality}
}
}
\source{
The \emph{Our World in Data} Dataset of Famines.
\url{https://ourworldindata.org/famines#the-our-world-in-data-dataset-of-famines}
}
\usage{
famines
}
\description{
Year, country, and mortality data for 77 famines.
}
\keyword{datasets}
