% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_combined.R
\docType{data}
\name{gm_combined}
\alias{gm_combined}
\title{Gapminder region, population, GDP, and life expectancy data}
\format{A data frame with 2562 rows and 7 variables.
\describe{
\item{iso_a3}{ISO alpha-3 country code}
\item{name}{Country name}
\item{region}{4-group Gapminder region}
\item{year}{Year}
\item{population}{Population}
\item{gdp_per_capita}{GDP per capita, PPP dollars}
\item{life_expectancy}{Life expectancy in years}
}}
\source{
Gapminder.
}
\usage{
gm_combined
}
\description{
Region, population, GDP, and life expectancy data for all countries from
1950 to 2015 every five years.
}
\keyword{datasets}
