% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_countries.R
\docType{data}
\name{gm_countries}
\alias{gm_countries}
\title{Gapminder country data}
\format{A data frame with 197 rows and 10 variables.
\describe{
\item{iso_a3}{ISO alpha-3 country code}
\item{name}{Country name}
\item{region_gm4}{4-group Gapminder region}
\item{region_gm6}{6-group Gapminder region}
\item{region_gm8}{8-group Gapminder region}
\item{region_wb}{World Bank region}
\item{oecd_g77}{Membership in OECD or Group of 77}
\item{un_status}{UN status}
\item{un_admission}{Date of UN admission}
\item{income_wb_2017}{4-group 2017 World Bank income group}
}}
\source{
Gapminder.
\url{https://www.gapminder.org/data/geo/}
}
\usage{
gm_countries
}
\description{
Basic country information from Gapminder
}
\keyword{datasets}
