% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_gdp.R
\docType{data}
\name{gm_gdp}
\alias{gm_gdp}
\title{Gapminder GDP per capita data}
\format{A data frame with 48734 rows and 4 variables.
\describe{
\item{iso_a3}{ISO alpha-3 country code}
\item{name}{Country name}
\item{year}{Year}
\item{gdp_per_capita}{GDP per capita, PPP dollars}
}}
\source{
Gapminder.
\url{https://www.gapminder.org/data/documentation/gd001/}
}
\usage{
gm_gdp
}
\description{
Historical and projected GDP per capita data for all countries from 1800 to
2040.
}
\keyword{datasets}
