% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_life_expectancy.R
\docType{data}
\name{gm_life_expectancy}
\alias{gm_life_expectancy}
\title{Gapminder life expectancy data}
\format{
A data frame with 60538 rows and 4 variables.
\describe{
\item{iso_a3}{ISO alpha-3 country code}
\item{name}{Country name}
\item{year}{Year}
\item{life_expectancy}{Life expectancy, in years}
}
}
\source{
Gapminder.
\url{https://www.gapminder.org/data/documentation/gd004/}
}
\usage{
gm_life_expectancy
}
\description{
Historical and projected life expectancy data for all countries from 1800 to
2100.
}
\keyword{datasets}
