% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_population.R
\docType{data}
\name{gm_population}
\alias{gm_population}
\title{Gapminder population data}
\format{A data frame with 59297 rows and 4 variables.
\describe{
\item{iso_a3}{ISO alpha-3 country code}
\item{name}{Country name}
\item{year}{Year}
\item{population}{Population}
}}
\source{
Gapminder.
\url{https://docs.google.com/spreadsheets/d/18Ep3s1S0cvlT1ovQG9KdipLEoQ1Ktz5LtTTQpDcWbX0/edit#gid=1668956939}
}
\usage{
gm_population
}
\description{
Historical and projected population data for all countries from 1800 to 2100.
}
\keyword{datasets}
