% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages_states.R
\docType{data}
\name{languages_states}
\alias{languages_states}
\title{Non-English languages spoken in US homes by state}
\format{A data frame with 5582 rows and 3 variables.
\describe{
\item{state}{State}
\item{language}{Language spoken at home}
\item{speakers}{Estimated number of speakers}
}}
\source{
2013 American Community Survey - Table Packages: Detailed Language
Spoken in the U.S.
\url{https://www.census.gov/data/developers/data-sets/language-stats.html}
}
\usage{
languages_states
}
\description{
Non-English languages spoken at home by the US population 5 years
and older in 2013, by state.
}
\keyword{datasets}
