% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages_us.R
\docType{data}
\name{languages_us}
\alias{languages_us}
\title{Non-English languages spoken in US homes in 2013}
\format{A data frame with 318 rows and 2 variables.
\describe{
\item{language}{Language spoken at home}
\item{speakers}{Estimated number of speakers}
}}
\source{
2013 American Community Survey - Table Packages: Detailed Language
Spoken in the U.S.
\url{https://www.census.gov/data/developers/data-sets/language-stats.html}
}
\usage{
languages_us
}
\description{
Non-English languages spoken at home by the US population 5 years
and older in 2013.
}
\keyword{datasets}
