% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_twins.R
\docType{data}
\name{example_twins}
\alias{example_twins}
\title{Example twin data}
\format{A data frame with 4 rows and 3 variables.
\describe{
\item{family}{Family name}
\item{name}{First name}
\item{n}{Birth order}
}}
\source{
XX
}
\usage{
example_twins
}
\description{
Birth-order data for two sets of twins.
}
\keyword{datasets}
